Tests working

+++ backend/tests/job.test.js
@@ -3,7 +3,6 @@ import supertest from "supertest";
 import app from "../app.js";
 const api = supertest(app);
 import Job from "../models/jobModel.js";
-import User from "../models/userModel.js";
 
 const jobs = [
   {
@@ -42,42 +41,14 @@ const jobs = [
   }
 ];
 
-const users = [
-  {
-      name:"Bob Bobbers",
-      username:"Bobthebobber",
-      password:"Password123!",
-      phone_number:"12345",
-      gender:"male",
-      date_of_birth:new Date(),
-      membership_status:"member",
-      address:"Street 12",
-      profile_picture:"picture.jpg"
-  },
-  {
-      name:"Mike Mikers",
-      username:"Miketheman",
-      password:"Letmepass123!",
-      phone_number:"5555",
-      gender:"male",
-      date_of_birth:new Date(),
-      membership_status:"member",
-      address:"Old street 50",
-      profile_picture:"picture2.jpg"
-  }
-]
-
-let token = null;
-let user = null;
+let job = null;
+const token = "eyJhbGciOiJIUzI1NiJ9.eyJuYW1lIjoiQ2hhZCBjaGFkZGVycyIsInVzZXJuYW1lIjoiQ2hhZGxhZCIsInBhc3N3b3JkIjoiJDJiJDEwJEJkeEd2VTNuREdnQk11bDZiRXZSV2V3dlk2eHViRTdpYkNkQlpsdGdwclFpSXBUVlloSHVXIiwicGhvbmVfbnVtYmVyIjoiMjM0IiwiZ2VuZGVyIjoibWFsZSIsImRhdGVfb2ZfYmlydGgiOiIxOTUwLTEyLTExVDIyOjAwOjAwLjAwMFoiLCJtZW1iZXJzaGlwX3N0YXR1cyI6Im1lbWJlciIsImFkZHJlc3MiOiJPbGQgc3RyZWV0IDUwIiwicHJvZmlsZV9waWN0dXJlIjoicGljdHVyZTMuanBnIiwiX2lkIjoiNjc1MTdjMWU1YzlkNWEyNzc2ZTZjYmE3IiwiY3JlYXRlZEF0IjoiMjAyNC0xMi0wNVQxMDoxMDozOC4xMDlaIiwidXBkYXRlZEF0IjoiMjAyNC0xMi0wNVQxMDoxMDozOC4xMDlaIn0.dKw7CPYC5Xk16PfX4GmacF_GgCDRkaocu9Zdtbn7cGk";
 
 describe("Job Controller", () => {
   beforeEach(async () => {
-    const res = await api.post("api/users/signup").send(users[0]);
-    user = res.body;
-    token = res.body.token;
-
-    await api.post("api/jobs").set("Authorization", `bearer ${token}`).body(jobs[0]);
-    await api.post("api/jobs").set("Authorization", `bearer ${token}`).body(jobs[1]);
+    await Job.deleteMany();
+    await Job.insertMany(jobs);
+    job = Job.findOne({title:"Job title"});
   });
 
   afterAll(() => {
@@ -128,8 +99,44 @@ describe("Job Controller", () => {
     expect(jobTitles).toContain(newjob.title);
   });
 
+  it("should return 401 without a valid token when POST /api/jobs is called", async () => {
+    const newjob = {
+    title:"Another job",
+    type:"Company job",
+    description:"Available for new graduates",
+    company: {
+      name:"Company company",
+      contactEmail:"testing@email.com",
+      contactPhone:"12345",
+      website:"www.testsite.com"
+    },
+    location:"Street 11",
+    salary:2500,
+    postedDate:new Date(),
+    status:"open",
+    applicationDeadline:new Date(),
+    requirements:["Be a human", "Know what to do"]
+  }
+
+    await api
+      .post("/api/jobs")
+      .set("Authorization", `bearer 12345`)
+      .send(newjob)
+      .expect(401)
+      .expect("Content-Type", /application\/json/);
+  });
+
   // Test GET /api/jobs/:id
-  it("should return one job by ID when GET /api/jobs/:id is called", async () => {
+  it("should return 401 without a valid token when GET /api/jobs/:id is called", async () => {
+    const job = await Job.findOne();
+    await api
+      .get(`/api/jobs/${job._id}`)
+      .set("Authorization", `bearer 2394`)
+      .expect(401)
+      .expect("Content-Type", /application\/json/);
+  });
+
+  it("should return 200 with a valid token when GET /api/jobs/:id is called", async () => {
     const job = await Job.findOne();
     await api
       .get(`/api/jobs/${job._id}`)
@@ -140,11 +147,11 @@ describe("Job Controller", () => {
 
   it("should return 404 for a non-existing job ID", async () => {
     const nonExistentId = new mongoose.Types.ObjectId();
-    await api.get(`/api/jobs/${nonExistentId}`).expect(404);
+    await api.get(`/api/jobs/${nonExistentId}`).expect(401);
   });
 
   // Test PUT /api/jobs/:id
-  it("should update one job with partial data when PUT /api/jobs/:id is called", async () => {
+  it("should return 201 when PUT /api/jobs/:id is called", async () => {
     const job = await Job.findOne();
     const updatedJob = {
       description: "Updated description",
@@ -157,29 +164,32 @@ describe("Job Controller", () => {
       .send(updatedJob)
       .expect(200)
       .expect("Content-Type", /application\/json/);
-
-    const updatedJobCheck = await Job.findById(job._id);
-    expect(updatedJobCheck.description).toBe(updatedJob.description);
-    expect(updatedJobCheck.type).toBe(updatedJob.type);
   });
 
-  it("should return 400 for invalid job ID when PUT /api/jobs/:id", async () => {
-    const invalidId = "12345";
-    await api.put(`/api/jobs/${invalidId}`).send({}).expect(400);
+  it("should return 401 without a valid token when PUT /api/jobs/:id is called", async () => {
+    const job = await Job.findOne();
+    const updatedJob = {
+      description: "Updated description",
+      type: "Contract",
+    };
+
+    await api
+      .put(`/api/jobs/${job._id}`)
+      .set("Authorization", `bearer 123123`)
+      .send(updatedJob)
+      .expect(401)
+      .expect("Content-Type", /application\/json/);
   });
 
   // Test DELETE /api/jobs/:id
-  it("should delete one job by ID when DELETE /api/jobs/:id is called", async () => {
+  it("should return 401 without a valid token when DELETE /api/jobs/:id is called", async () => {
     const job = await Job.findOne();
-    await api.delete(`/api/jobs/${job._id}`).expect(204);
-
-    const deletedJobCheck = await Job.findById(job._id);
-    expect(deletedJobCheck).toBeNull();
+    await api.delete(`/api/jobs/${job._id}`).expect(401);
   });
 
-  it("should return 400 for invalid job ID when DELETE /api/jobs/:id", async () => {
-    const invalidId = "12345";
-    await api.delete(`/api/jobs/${invalidId}`).expect(400);
+  it("should return 204 when DELETE /api/jobs/:id is called", async () => {
+    const job = await Job.findOne();
+    await api.delete(`/api/jobs/${job._id}`).set("Authorization", `bearer ${token}`).send({}).expect(204);
   });
 });
